feat: report button on bounty item page
This PR adds a report system and a button on the bounty item page. The report will publish a 1984 event, which will then be filtered out of the main bounty events cache if the filter rules pass. In this case, if a bounty has one report, it will be filtered out.